// packages/ioc/src/injector.d.ts
export declare const RECORDS: unique symbol;
/**
 * injector.
 * implements {@link Destroyable}
 *
 * IoC 容器，注入器
 */
export declare abstract class Injector implements Destroyable, OnDestroy {
    /**
     * 是否静态容器。
     */
    readonly isStatic?: boolean;
    /**
     * records of providers.
     *
     * 容器提供者记录
     */
    abstract readonly [RECORDS]: Map<Token<any>, InjectorRecord>;
    /**
     * injector scope.
     *
     * 容器范围
     */
    readonly scope?: InjectorScope;
    /**
     * init inject ready.
     */
    abstract get ready(): Promise<void>;
    /**
     * parent injector.
     *
     * 上级容器。
     */
    abstract getParent(): Injector;
    /**
     * get runtime.
     *
     * 容器运行环境
     */
    abstract getRuntime(): Runtime;
    /**
     * has register.
     *
     * 标记令牌是否已注册。
     *
     * @template T
     * @param {Token<T>} token the token.
     * @param {InjectFlags} flags check strategy by inject flags {@link InjectFlags}.
     * @returns {boolean}
     */
    abstract has<T>(token: Token<T>, flags?: InjectFlags): boolean;
    /**
     * get token factory resolve instace in current.
     *
     * 获取标记令牌的实例。
     * @template T
     * @param {Token<T>} token token id {@link Token}.
     * @param {T} notFoundValue not found token, return this value.
     * @param {InjectFlags} flags check strategy by inject flags {@link InjectFlags}.
     * @param {RunContext} context resolve context. type of {@link RunContext}, use to resolve with token.
     * @returns {T} token value.
     */
    abstract get<T>(token: Token<T>, notFoundValue?: T, flags?: InjectFlags, context?: RunContext): T;
    /**
     * resolve parameter of targetType.
     *
     * 解析标记令牌的实例。
     *
     * @template T
     * @param {Parameter<T>} parameter the resolve parameter {@link Parameter}.
     * @param {RunContext} context the resolver context.
     *
     * @returns {T}
     */
    abstract resolve<T>(parameter: Parameter<T>, context?: RunContext): T;
    /**
     * resolve token in context.
     *
     * 解析上下文中标记指令的实例值
     * @param token
     * @param {RunContext} context the resolver context.
     */
    abstract resolve<T>(token: Token<T>, context?: RunContext): T;
    /**
     * resolve token in context.
     *
     * 解析上下文中标记指令的实例值
     * @param token
     * @param flags InjectFalgs
     */
    abstract resolve<T>(token: Token<T>, falgs?: InjectFlags, context?: RunContext): T;
    /**
     * set value.
     *
     * 设置上下文中标记指令的实例值
     * @param token token
     * @param value value for the token.
     */
    abstract setValue<T>(token: Token<T>, value: T): this;
    /**
     * injector has destoryed or not.
     */
    abstract get destroyed(): boolean;
    /**
    * destroy this.
    */
    abstract destroy(): void;
    /**
     * destroy hook.
     */
    abstract onDestroy(): void;
    /**
     * register callback on destroy.
     * @param callback destroy callback
     */
    abstract onDestroy(callback: DestroyCallback): void;
}
export declare abstract class EnvironmentInjector extends Injector {
}
// ...
export declare const INJECT_IMPL: {
    createRoot(providers?: Provider[]): EnvironmentInjector;
    /**
     * create injector
     * @param parent
     * @param providers
     * @param scope
     */
    create(parent: Injector, providers?: Provider[], scope?: AbstractType | "root" | "static"): Injector;
    /**
     * create injector
     * @param parent
     * @param options
     * @param scope
     */
    createByOptions(parent: Injector, options?: InvokeProviders, scope?: AbstractType | "static"): Injector;
    isInjector(target: any): boolean;
};
/**
 * instance factory.
 */
export type Factory<T = any> = (...args: any[]) => T;
export type Resolve<T> = (injector: Injector) => T | null | undefined;
export type InstanceOf<T> = T | Resolve<T>;
/**
 * register option
 */
export interface RegOption<T = any> extends ProvidedInMetadata {
    provide?: Token<T>;
    injectorType?: boolean;
    regProvides?: boolean;
    singleton?: boolean;
    static?: boolean;
    declaration?: boolean;
}
/**
 * type register option.
 */
export interface TypeOption<T = any> extends RegOption<T> {
    type: AbstractType<T>;
}
/**
 * provider option.
 */
export type ProviderOption<T = any> = ClassProvider<T> | ValueProvider<T> | ExistingProvider<T> | FactoryProvider<T>;
/**
 * register option.
 */
export type RegisterOption<T = any> = TypeOption<T> | ProviderOption<T>;
/**
 * injector scope.
 */
export type InjectorScope = AbstractType | 'platform' | 'root' | 'static';
export interface MethodFunc extends Function, TypedPropertyDescriptor<any> {
}
/**
 * method type.
 */
export type MethodType<T> = string | symbol | ((tag: T) => MethodFunc);
export type RecordFactory<T = any> = (context?: RunContext, flags?: InjectFlags) => T | null;
/**
 * Injecor Record
 */
export interface InjectorRecord<T = any> {
    /**
     * 提供者的类型
     */
    type?: AbstractType<T>;
    /**
     * 工厂函数，用于创建实例
     */
    factory?: RecordFactory<T>;
    /**
     * 预创建的值或占位符
     */
    value?: T | null | {};
    /**
     * 多提供者的依赖数组
     */
    multi?: any[];
    /**
     * is static or not.
     */
    stati?: boolean;
    /**
     * cache expires.
     */
    expires?: number;
    onRegister?: () => void;
}